Rename service-worker to service-worker-manager
