+++ website/src/views/settings/RefreshPrompt.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import classnames from 'classnames';
 import { connect } from 'react-redux';
 
-import { updateServiceWorker } from 'bootstrapping/service-worker';
+import { updateServiceWorker } from 'bootstrapping/service-worker-manager';
 import { Refresh } from 'views/components/icons';
 import styles from 'views/components/notfications/Announcements.scss';
 import LoadingSpinner from 'views/components/LoadingSpinner';
